+++ registry/components/chart/index.ts
@@ -1,0 +1,3 @@
+export { default as LineChart } from "./line-chart.svelte";
+export { default as BarChart } from "./bar-chart.svelte";
+export type { Granularity } from "./format.ts";