+++ registry/components/navbar/index.ts
@@ -1,0 +1,3 @@
+export { default as Navbar } from "./navbar.svelte";
+export { default as NavbarLink } from "./navbar-link.svelte";
+export { default as NavbarMobileMenu } from "./navbar-mobile-menu.svelte";