+++ registry/components/sidebar/index.ts
@@ -1,0 +1,7 @@
+export { default as Sidebar } from "./sidebar.svelte";
+export { default as SidebarHeader } from "./sidebar-header.svelte";
+export { default as SidebarContent } from "./sidebar-content.svelte";
+export { default as SidebarFooter } from "./sidebar-footer.svelte";
+export { default as SidebarGroup } from "./sidebar-group.svelte";
+export { default as SidebarMenu } from "./sidebar-menu.svelte";
+export { default as SidebarMenuItem } from "./sidebar-menu-item.svelte";